+++ CHANGES.md
@@ -23,6 +23,7 @@ Change Log
 * Fixed `ModelNode.matrix` when a node is targeted for animation.
 * `Camera.tilt` now clamps to [-pi / 2, pi / 2] instead of [0, pi / 2].
 * Fixed an issue that could lead to poor performance on lower-end GPUs like the Intel HD 3000.
+* Added `distanceSquared` to `Cartesian2`, `Cartesian3`, and `Cartesian4`.
 
 ### 1.3 - 2014-11-03